requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.2.2
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
